requirements.txt
attrs==20.2.0
aws-cdk.assets==1.69.0
aws-cdk.aws-apigateway==1.69.0
aws-cdk.aws-applicationautoscaling==1.69.0
aws-cdk.aws-autoscaling-common==1.69.0
aws-cdk.aws-certificatemanager==1.69.0
aws-cdk.aws-cloudwatch==1.69.0
aws-cdk.aws-codeguruprofiler==1.69.0
aws-cdk.aws-ec2==1.69.0
aws-cdk.aws-efs==1.69.0
aws-cdk.aws-elasticloadbalancingv2==1.69.0
aws-cdk.aws-events==1.69.0
aws-cdk.aws-iam==1.69.0
aws-cdk.aws-kms==1.69.0
aws-cdk.aws-lambda==1.69.0
aws-cdk.aws-logs==1.69.0
aws-cdk.aws-route53==1.69.0
aws-cdk.aws-s3==1.69.0
aws-cdk.aws-s3-assets==1.69.0
aws-cdk.aws-sqs==1.69.0
aws-cdk.aws-ssm==1.69.0
aws-cdk.cloud-assembly-schema==1.69.0
aws-cdk.core==1.69.0
aws-cdk.cx-api==1.69.0
aws-cdk.region-info==1.69.0
aws-dev-env==0.0.1
cattrs==1.0.0
constructs==3.0.27
jsii==1.13.0
publication==0.0.3
python-dateutil==2.8.1
PyYAML==5.3.1
six==1.15.0
typing-extensions==3.7.4.3
